#105 - Cria useQuery para o endpoint de project
