+++ src/pages/projects/[project]/components/ProjectContent/ProjectContent.tsx
@@ -1,45 +1,17 @@
-import React, { useEffect, useState } from 'react';
-
-import { useRouter } from 'next/router';
+import React from 'react';
 
 import { formatRelative } from 'date-fns';
 import { ptBR } from 'date-fns/locale';
 
 import { Box, Typography } from '@mui/material';
 
-import { projectQuery } from '@services/index';
-
 import Skeleton from './Skeleton';
 
 import * as Styles from './styles';
-
-interface Project {
-  id: number;
-  name: string;
-  description: string;
-  github_url: string;
-  created_at: string;
-  updated_at: string;
-}
+import useQuery from './hook/useQuery';
 
 const ProjectContent = () => {
-  const { query } = useRouter();
-  const [project, setProject] = useState<Project>();
-
-  async function loadQuery() {
-    if (query?.project) {
-      try {
-        const result = await projectQuery.getProjectById('1', query?.project as string);
-        setProject(result.data);
-      } catch (error) {
-        return error;
-      }
-    }
-  }
-
-  useEffect(() => {
-    loadQuery();
-  }, [query?.project]);
+  const { project } = useQuery();
 
   const lastUpdateDate =
     project &&
